Exposing struct_utils for accessing custom JWT fields
Exposing struct_utils to enable custom JWT fields to be retrieved from a
parsed JWT using the existing `payload_pb` field
